Refactoring and add test for new command
